Split tty and test tty C files
